Unit testing code shifted to the seperate file by Sayma
